requirements.txt
google-genai==1.69.0
python-dotenv==1.2.2